rename(store): dto 디렉토리 분리
