--- src/main/java/com/example/springrider/domain/store/dto/UpdateStoreRequestDto.java
@@ -1,37 +1,0 @@
-package com.example.springrider.domain.store.dto;
-
-import com.example.springrider.domain.store.enums.StoreStatus;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import java.time.LocalTime;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(force = true)
-public class UpdateStoreRequestDto {
-
-    @NotBlank(message = "가게 이름은 필수입니다.")
-    private String name;
-
-    @NotBlank(message = "가게 주소는 필수입니다.")
-    private String address;
-
-    @NotBlank(message = "음식 장르는 필수입니다.")
-    private String category;
-
-    @NotNull(message = "오픈 시간은 필수입니다.")
-    private LocalTime openTime;
-
-    @NotNull(message = "마감 시간은 필수입니다.")
-    private LocalTime closeTime;
-
-    @Positive(message = "최소 주문 금액은 1원 이상이어야 합니다.")
-    private Integer minOrderPrice;
-
-    private StoreStatus status;
-
-}
