+++ src/main/java/com/example/springrider/domain/store/entity/Store.java
@@ -3,12 +3,12 @@ package com.example.springrider.domain.store.entity;
 import static com.example.springrider.global.exception.ExceptionCode.STORE_INVALID_STATUS_CHANGE;
 
 import com.example.springrider.domain.common.entity.BaseEntity;
-import com.example.springrider.global.exception.InvalidRequestException;
 import com.example.springrider.domain.menu.entity.Menu;
-import com.example.springrider.domain.store.dto.StoreRequestDto;
-import com.example.springrider.domain.store.dto.UpdateStoreRequestDto;
+import com.example.springrider.domain.store.dto.request.StoreRequestDto;
+import com.example.springrider.domain.store.dto.request.UpdateStoreRequestDto;
 import com.example.springrider.domain.store.enums.StoreStatus;
 import com.example.springrider.domain.user.entity.User;
+import com.example.springrider.global.exception.InvalidRequestException;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
