+++ src/test/java/com/example/springrider/domain/store/service/StoreServiceTest.java
@@ -9,18 +9,18 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.example.springrider.global.exception.InvalidRequestException;
-import com.example.springrider.domain.store.dto.FindAllStoreResponseDto;
-import com.example.springrider.domain.store.dto.FindStoreResponseDto;
-import com.example.springrider.domain.store.dto.StoreRequestDto;
-import com.example.springrider.domain.store.dto.StoreResponseDto;
-import com.example.springrider.domain.store.dto.UpdateStoreRequestDto;
+import com.example.springrider.domain.store.dto.request.StoreRequestDto;
+import com.example.springrider.domain.store.dto.request.UpdateStoreRequestDto;
+import com.example.springrider.domain.store.dto.response.FindAllStoreResponseDto;
+import com.example.springrider.domain.store.dto.response.FindStoreResponseDto;
+import com.example.springrider.domain.store.dto.response.StoreResponseDto;
 import com.example.springrider.domain.store.entity.Store;
 import com.example.springrider.domain.store.enums.StoreStatus;
 import com.example.springrider.domain.store.repository.StoreRepository;
 import com.example.springrider.domain.user.entity.User;
 import com.example.springrider.domain.user.enums.UserRole;
 import com.example.springrider.domain.user.repository.UserRepository;
+import com.example.springrider.global.exception.InvalidRequestException;
 import java.lang.reflect.Field;
 import java.time.LocalTime;
 import java.util.List;
